testing index_unidim_to_multidim()

+++ plugins/plugin_utils.c
@@ -101,7 +101,7 @@ int32_t deserialize_meta(uint8_t *smeta, int32_t smeta_len, int8_t *ndim, int64_
   return slen;
 }
 
-void index_unidim_to_multidim(uint8_t ndim, int32_t *shape, int64_t i, int64_t *index) {
+void index_unidim_to_multidim(uint8_t ndim, int64_t *shape, int64_t i, int64_t *index) {
     int64_t strides[8];
     strides[ndim - 1] = 1;
     for (int j = ndim - 2; j >= 0; --j) {
